core/requirements.txt
asgiref==3.8.1
coverage==7.6.1
Django==4.2.16
django-environ==0.11.2
sqlparse==0.5.1
